backend/pyproject.toml
[project]
name = "tiramai-rag-backend"
version = "1.0.0"
description = "Tiramai RAG System Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.1",
    "chromadb>=0.4.0",
    "docling>=1.0.0",
    "httpx>=0.25.2",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "memorisdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

